[REP-114 + REP-83 -> master] Set snapshot status when initiating a delete and when process is complete

+++ db/migrations/20180226093552_add_receive_failed_status_to_snapshot_status.js
@@ -1,29 +1,21 @@
 
 exports.up = async function(knex) {
-    await knex.transaction(async (tx) => {
-        try {
+    knex.schema.hasTable('snapshot_status').then(async (exists) => {
+        if (exists) {
             await knex('snapshot_status').insert([{
                 id: 4,
                 name: 'receive_failed'
             }]);
-
-            await tx.commit();
-        } catch (err) {
-            console.log(err);
-            await tx.rollback();
         }
     });
 };
 
 exports.down = async function(knex) {
-    await knex.transaction(async (tx) => {
-        try {
+    knex.schema.hasTable('snapshot_status').then(async (exists) => {
+        if (exists) {
             await knex('snapshot_status').where('id', 4).delete();
-
-            await tx.commit();
-        } catch (err) {
-            console.log(err);
-            await tx.rollback();
         }
     });
 };
+
+exports.config = { transaction: true };
